feat(BonusScore): extract interface
- 보너스 점수의 인터페이스 추출
