+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -1,6 +1,6 @@
 package bowling.domain.frame;
 
-import bowling.domain.BonusScore;
+import bowling.domain.NormalBonusScore;
 import bowling.domain.FrameResults;
 import bowling.domain.FrameScore;
 import bowling.domain.FrameScoreStatus;
@@ -47,7 +47,7 @@ public class NormalFrame implements Frame {
         }
     }
 
-    private BonusScore calculateBonusScore() {
+    private NormalBonusScore calculateBonusScore() {
         return this.currentStatus.calculateBonusScore();
     }
